RTC register address is all changed.
RTC clock configuration has a mistake in the xsysctl.c line 1057.I modified as follows:
original:
   xHWREG(RCC_BDCR) &= ~(RCC_BDCR_RTCSEL_M);
now:
   xHWREG(RCC_BDCR) |=(((ulPeripheralSrc)&0xf)<<8);

The name of interrupt handling function have a underline in the start-up code of keil, not compatible with CoX, need to be careful and change.

the RTC counter register(RTC_CNTH) can not be write and read the correct value.The line 402、408、473、478 of xrtc.c must be modified.